Rework to define piece using enum type instead of usize
For now this is done in new module pieces2.rs, and pieces.rs is
retained in case there are some definitions there still to be
transferred.  Modules that use the code have been updated to "use
pieces2 as pieces".

In general, this approach seems to make more sense and be simpler in
many cases, although I'm not 100% sure of the performance
implications.  Most of the "accessors" are now methods instead of
arrays.  There are still some cases that require casting to usize to
perform indexing on the piece.

The Move struct has also been reworked and dramatically simplified.
Instead of packing the data into bits, it is now laid out as regular
struct fields.  This results in requiring 4 bytes instead of 5 bytes
for move data, but the access is much simpler.