Add type alias Square
And define it as i8.  This provides a single definition of the Square
type, making the type definitions more clear.  Also, since it is
signed, it eliminates the need to do casting when adding directions.
The drawback is that casting is needed when doing array indexing, but
that seems to be more consistent with the conceptual "square" type
anyway.

Another advantage is that squares are now represented using only 1
byte, whereas previously 8 bytes were being used (usize).