Change MoveList struct to contain vec of ScoredMove
Remove the score value from Move and create new struct ScoredMove that
contains a Move and a score.  This parallels what is done in VICE.  In
this context, advantage seems to be that we can ensure that when
working with ScoredMove items, we always have a score that is defined.