Start of UCI implementation, VICE video 68
Added Board.update_from_fen, which takes in a FEN string and returns a
new board.  The difference from Board::from_fen is that it retains the
pv_table.  Doing it this way seemed preferable to the code duplication
that would be involved with having a separate reset_board function.

Implemented uci_loop function, as well as the parse_pos function.
Next up is to implement the parse_go function.