Add capability for search to detect "stop" and "quit" on stdin
The UCI loop sets up a separate thread for reading stdin.  This thread
then stores a reference to the receiver in SearchInfo, and the checkup
function can check this for pending input.

I have noticed that the program essentially floors one CPU now just
when waiting for input, which it did not do before.  Should review
whether there is a way to address this.