Handle xboard "remove" command
This is used when the user asks to undo a move.  Simply undo the last
two moves and continue playing same color.  From initial testing, this
does seem to be working (previously, it would get obviously hung up if
the user undid a move and then played it again, as chareth thought it
was illegal the second time because it hadn't actually undone the
move).