Overhaul of move time allocation
Moving the calculations out of the UCI and xboard loops and into new
function that is a member of SearchInfo.  Change to how the increment
is handled.  Previously, following VICE, the increment was added to
the computed move time.  However, that can result in over-running the
clock if the increment is large relative to the remaining clock time.
With this update, we compute a total remaining time with increments
added, and then divide that by the number of moves.  Checks have also
been added to ensure a minimum move time (just to make sure we can at
least find a move.