Add in_check extension
Based on VICE video 76, but doesn't exactly follow that
implementation.  I have placed it prior to the quiescence search
instead of after (as in VICE and tscp).  It makes more sense to me to
avoid going into quiescence search while in check.

Node counts for both of the two search test cases needed to be
increased.