Set up bitboard pawn evaluation masks
Following VICE video 78.  Believe that the initializations are
correct, although they could be reviewed more closely, as there is
some tricky code.

Currently using static mut globals, but it is a little cumbersome
because it requires unsafe code.  Believe the only alternative would
be to package the data that need to be initialized into a struct,
instantiate it in main, and then pass around a reference where it is
needed.