Switch to bitboard move generation for king and knight
Yields about 2% speedup in search.  More should be possible after a
full transition to bitboards to eliminate the need for 64-120
conversion.

Resulted in a small increase in node count for search_wac1_depth3
test (and noticed similar increase in depth-9 starting position
search).  Maybe has to do somehow with slight difference in the order
that the move directions are checked?

Verified that full perftsuite test still passes.