Switch to 64-square representation
120-square representation no longer needed now that everything is done
with bitboards.  For now, Square is defined as usize to avoid need to
cast with array access, but may be better to switch over to u8 to save
space.  There is probably some other cleanup as well that could be
done, in particular to clarify some of the conversions such as ranks,
files, and direction-based en passant calculations.

Yields about a 2% speedup in search.